Reward's can now accept decimals
- Update IGC to allow double values for reward chance
- Config file now supports loading double values for reward chance
- Update random reward system to use decimal chances

+++ src/main/java/me/ztowne13/customcrates/crates/options/rewards/Reward.java
@@ -291,7 +291,7 @@ public class Reward
     {
         try
         {
-            setChance(getCc().getRewardsFile().get().getInt(getPath("chance")));
+            setChance(getCc().getRewardsFile().get().getDouble(getPath("chance")));
         }
         catch (Exception exc)
         {
